requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
flask_smorest==0.42.0
yfinance==0.2.28
openai==0.28.0
requests==2.31.0
pandas==1.4.2
python-dotenv==0.20.0